requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
django-cors-headers==3.13.0
Django==4.0.5
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
Faker==13.14.0
pip==20.0.2
pkg-resources==0.0.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
setuptools==44.0.0
six==1.16.0
sqlparse==0.4.2
wheel==0.34.2
